Fixed warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_datatools/trunk@155 f95d2973-9550-45a5-b795-f9f66c6b6f94

+++ src/io/MMGDDDataSource.cpp
@@ -129,7 +129,7 @@ bool io::MMGDDDataSource::filenameChanged(core::param::ParamSlot& slot) {
     double memHere = static_cast<double>(vislib::sys::SystemInformation::AvailableMemorySize());
     memHere *= 0.25; // only use max 25% of the memory of this data
     Log::DefaultLog.WriteInfo("Memory available: %u MB\n", static_cast<uint32_t>(memHere / (1024.0 * 1024.0)));
-    double memWant = frameIdx.back() - frameIdx.front();
+    double memWant = static_cast<double>(frameIdx.back() - frameIdx.front());
     Log::DefaultLog.WriteInfo("Memory required: %u MB for %u frames total\n", static_cast<uint32_t>(memWant / (1024.0 * 1024.0)), static_cast<uint32_t>(frameIdx.size()));
     uint32_t cacheSize = static_cast<uint32_t>((memHere / memWant) * static_cast<double>(frameIdx.size()) + 0.5);
     Log::DefaultLog.WriteInfo("Cache set to %u frames\n", cacheSize);
